requirements.txt
certifi==2020.6.20
glob2==0.7
numpy==1.19.4
Pillow==8.0.1
tqdm==4.52.0
